* unsigned int changed into uint8_t for memory saving
Now path seeking works even for Eulerian graphs.

+++ dfs.h
@@ -9,12 +9,12 @@
 #ifndef _DFS_H_
 #define _DFS_H_
 
-#include <stdbool.h>
+#include <stdint.h>
 
 typedef struct graphParam{
     size_t depth;                   // Max depth I can go into
-    unsigned int **ops;             // Set of operations
-    unsigned int origins[2];        // Set of origin nodes
+    uint8_t **ops;             // Set of operations
+    uint8_t origins[2];        // Set of origin nodes
 } GParams;
 
 void warmUp(GParams *details);      // Call this and start seeking
